@@ -8,43 +8,47 @@
     "url": "https://ztrehagem.dev/"
   },
   "scripts": {
-    "dev": "vue-cli-service serve",
-    "build": "run-s build:*",
-    "build:lib": "vue-cli-service build --target=lib --name=vue-modal ./src/lib/main.ts",
-    "build:dts": "tsc -p tsconfig.build.json",
-    "gh-pages": "vue-cli-service build --dest=gh-pages && touch gh-pages/.nojekyll",
+    "dev": "vite",
+    "build": "vite build && tsc -p tsconfig.dts.json",
     "lint": "run-s lint:*",
     "lint:eslint": "eslint --ignore-path=.gitignore .",
     "lint:prettier": "prettier --ignore-path=.gitignore --check .",
     "fix": "run-s fix:*",
     "fix:eslint": "eslint --ignore-path=.gitignore --fix .",
     "fix:prettier": "prettier --ignore-path=.gitignore --check --write .",
-    "prepare": "run-s build"
+    "prepack": "run-s build"
   },
-  "main": "dist/vue-modal.common.js",
+  "main": "dist/vue-modal.cjs.js",
+  "module": "dist/vue-modal.es.js",
+  "exports": {
+    ".": {
+      "import": "./dist/vue-modal.es.js",
+      "require": "./dist/vue-modal.cjs.js"
+    },
+    "./style.css": "./dist/style.css"
+  },
+  "types": "./dist/types/main.d.ts",
   "files": [
     "dist"
   ],
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-typescript": "^4.5.15",
-    "@vue/cli-service": "~4.5.0",
+    "@babel/types": "^7.17.0",
+    "@types/node": "^17.0.17",
+    "@vitejs/plugin-vue": "^2.2.0",
+    "@vue/eslint-config-prettier": "^7.0.0",
     "@vue/eslint-config-typescript": "^9.1.0",
-    "core-js": "^3.6.5",
     "eslint": "^8.3.0",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier-vue": "^3.1.0",
     "eslint-plugin-vue": "^8.1.1",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.5.0",
     "sass": "^1.43.4",
-    "sass-loader": "^10.2.0",
-    "typescript": "~4.5.2",
-    "vue": "2.6.0",
-    "vue-template-compiler": "2.6.0"
+    "typescript": "^4.5.5",
+    "vite": "^2.8.1",
+    "vue": "^3.2.31"
   },
   "peerDependencies": {
-    "vue": "^2.6.0"
+    "vue": "^3.0.0"
   },
   "bugs": "https://github.com/ztrehagem/vue-modal/issues",
   "homepage": "https://github.com/ztrehagem/vue-modal",
@@ -59,6 +63,5 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/ztrehagem/vue-modal.git"
-  },
-  "typings": "dist/main.d.ts"
+  }
 }
