@@ -1,3 +1,17 @@
+<script setup lang="ts">
+import { ref } from "vue";
+import { VueModalBackdrop, VueModalRenderer } from "./lib/main";
+import { useModal } from "./plugins/modal";
+
+const modal = useModal();
+
+const name = ref("");
+
+function showModal() {
+  modal.push("hello", { name: name.value });
+}
+</script>
+
 <template>
   <div id="app">
     <div>
@@ -8,28 +22,3 @@
     <VueModalRenderer />
   </div>
 </template>
-
-<script lang="ts">
-import Vue from "vue";
-import { VueModalBackdrop } from "@/lib/main";
-import { VueModalRenderer } from "@/lib/main";
-
-export default Vue.extend({
-  components: {
-    VueModalBackdrop,
-    VueModalRenderer,
-  },
-
-  data() {
-    return {
-      name: "",
-    };
-  },
-
-  methods: {
-    showModal() {
-      this.$modal.push("hello", { name: this.name });
-    },
-  },
-});
-</script>
