@@ -1,10 +1,11 @@
-import plugin from "./plugin";
 import VueModal from "./VueModal.vue";
 import VueModalBackdrop from "./VueModalBackdrop.vue";
 import VueModalRenderer from "./VueModalRenderer.vue";
 
-export * from "./ModalManager";
+export {
+  ModalManager,
+  ModalManagerInjectionError,
+  useModal,
+} from "./ModalManager";
 
-export { plugin, VueModal, VueModalBackdrop, VueModalRenderer };
-
-export default plugin;
+export { VueModal, VueModalBackdrop, VueModalRenderer };
